Consolidate TablesRegistrySchemaTest into DataConnectE2eTest
